Object update
objects can be picked up, or taken into inventory. once there they can be equipped which will grant the player access to the object's functions. This can also be accomplished by picking up an item. Added functions to remove a certain object in the player's room (to be used in gameplay) and to remove an object in a certain area (this is a more general feature to use when editing a world).